+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Package validation is disabled by default. [#] (https://github.com/elastic/package-registry/pull/)
+
 ### Deprecated
 
 ### Known Issues